requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
isort==5.13.2
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.14.1
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2
pip~=24.1.1
attrs~=22.2.0
wheel~=0.43.0
Pillow~=9.3.0
tomli~=2.0.1
colorama~=0.4.6
setuptools~=57.5.0
telebot~=0.0.5